requirements.txt
# Minimal requirements for Jira fetcher - Iterative Approach
# Python 3.10 - 3.13 compatible

# Jira API Client
jira>=3.5.0

# Data Analysis
polars>=0.20.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Data Validation
pydantic>=2.0.0

# Web Framework (for FastAPI service)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
jinja2>=3.1.3
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
